Implementing multiple requirements authorization handler with object? resource parameter

+++ Toyer/Toyer.API/Controllers/UserController.cs
@@ -1,6 +1,9 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Toyer.Data.Entities;
 using Toyer.Logic.Dtos.User;
+using Toyer.Logic.Exceptions.FailResponses.Derived.User;
 using Toyer.Logic.Responses;
 using Toyer.Logic.Services.Repositories.Interfaces;
 
@@ -8,12 +11,18 @@ namespace Toyer.API.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
+[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
 [Produces("application/json")]
 [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-public class UserController(IUserRepository userRepository, IUserMapings mappings, IDeviceAssignRepository deviceAssignRepository) : ControllerBase
+public class UserController(IUserRepository userRepository, 
+    IUserMapings mappings, 
+    IDeviceAssignRepository deviceAssignRepository, 
+    IAuthorizationService authorizationService) 
+    : ControllerBase
 {
     private readonly IUserMapings _mappings = mappings;
     private readonly IDeviceAssignRepository _deviceAssignRepository = deviceAssignRepository;
+    private readonly IAuthorizationService _authorizationService = authorizationService;
     private readonly IUserRepository _userRepository = userRepository;
 
 
@@ -21,11 +30,12 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     /// Logs user in.
     /// </summary>
     [HttpPost("Login")]
+    [AllowAnonymous]
     public async Task<IActionResult> Authenticate([FromForm] UserLogin request)
     {
         var result = await _userRepository.LoginAsync(request.Email, request.Password);
 
-        return new ObjectResult (result);
+        return new ObjectResult(result);
     }
 
 
@@ -45,6 +55,7 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     /// Gets all users.
     /// </summary>
     [HttpGet]
+    [Authorize(Policy = "ProductionTasks")]
     [ProducesResponseType(typeof(IEnumerable<UserPresentShortDto>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status202Accepted)]
     [ProducesResponseType(typeof(UnauthorizedResult), StatusCodes.Status401Unauthorized)]
@@ -63,6 +74,9 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status404NotFound)]
     public async Task<IActionResult> GetExtendedById([FromRoute] string userId)
     {
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, userId, "OwnerOrPrivilegedPolicy");
+        if (!authorizationResult.Succeeded) throw new AccessException();
+
         var user = await _userRepository.GetUserByIdAsync(userId);
 
         return Ok(_mappings.UserToUserPresentLongDto(user));
@@ -72,6 +86,7 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     /// Creates account for new user
     /// </summary>
     [HttpPost]
+    [AllowAnonymous]
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status201Created)]
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> CreateNewUser([FromForm] UserCreateDto newUserDto)
@@ -79,7 +94,6 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
         var result = await _userRepository.RegisterNewUserAsync(_mappings.UserCreateDtoToUser(newUserDto), newUserDto.Password);
 
         return CreatedAtAction(nameof(CreateNewUser), new CustomResponse { Message = $"User: {newUserDto.UserName} created.", StatusCode = 201 });
-            
     }
 
     /// <summary>
@@ -90,8 +104,10 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status404NotFound)]
     public async Task<IActionResult> UpdateAddressById([FromRoute] string userId, [FromForm] AddressDto addressUpdatesDtoFromUser)
     {
-        await _userRepository.UpdateAddressAsync(userId, _mappings.AddressDtoToAddress(addressUpdatesDtoFromUser));
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, userId, "OwnerOrPrivilegedPolicy");
+        if (!authorizationResult.Succeeded) throw new AccessException();
 
+        await _userRepository.UpdateAddressAsync(userId, _mappings.AddressDtoToAddress(addressUpdatesDtoFromUser));
         return NoContent();
     }
 
@@ -103,8 +119,10 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status404NotFound)]
     public async Task<IActionResult> UpdatePersonalInfoById([FromRoute] string userId, [FromForm] PersonalInfoDto personalInfoUpdates)
     {
-         await _userRepository.UpdatePersonalInfoPatchAsync(userId, _mappings.PersonalInfoDtoToPersonalInfo(personalInfoUpdates));
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, userId, "OwnerOrPrivilegedPolicy");
+        if (!authorizationResult.Succeeded) throw new AccessException();
 
+        await _userRepository.UpdatePersonalInfoPatchAsync(userId, _mappings.PersonalInfoDtoToPersonalInfo(personalInfoUpdates));
         return NoContent();
     }
 
@@ -117,8 +135,10 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> UpdateContactInfoById([FromRoute] string userId, [FromForm] ContactDto contactUpdates)
     {
-        await _userRepository.UpdateContactInfoAsync(userId, contactUpdates.Email, contactUpdates.PhoneNumber);
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, userId, "OwnerOrPrivilegedPolicy");
+        if (!authorizationResult.Succeeded) throw new AccessException();
 
+        await _userRepository.UpdateContactInfoAsync(userId, contactUpdates.Email, contactUpdates.PhoneNumber);
         return NoContent();
     }
 
@@ -131,10 +151,12 @@ public class UserController(IUserRepository userRepository, IUserMapings mapping
     [ProducesResponseType(typeof(CustomResponse), StatusCodes.Status400BadRequest)]
     public async Task<ActionResult> DeleteUserById([FromRoute] string userId)
     {
+
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, userId, "OwnerOrPrivilegedPolicy");
+        if (!authorizationResult.Succeeded) throw new AccessException();
+
         await _userRepository.DeleteUserAsync(userId);
-
         return NoContent();
-
     }
